refactor: 메뉴그룹 관련 리팩터링
- jpa 디펜던시 추가
- 패키지 경로 추가
- set 메서드 제거
