+++ src/main/java/kitchenpos/menugroup/domain/MenuGroup.java
@@ -1,9 +1,14 @@
-package kitchenpos.domain;
+package kitchenpos.menugroup.domain;
 
+import javax.persistence.*;
 import java.util.Objects;
 
+@Entity
 public class MenuGroup {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+    @Column(unique = true)
     private String name;
 
     public MenuGroup() {
@@ -27,16 +32,8 @@ public class MenuGroup {
         return id;
     }
 
-    public void setId(final Long id) {
-        this.id = id;
-    }
-
     public String getName() {
         return name;
-    }
-
-    public void setName(final String name) {
-        this.name = name;
     }
 
     @Override
