+++ src/main/resources/application.properties
@@ -1,2 +1,4 @@
 logging.level.org.springframework.jdbc.core=TRACE
 spring.h2.console.enabled=true
+
+spring.jpa.hibernate.ddl-auto=validate
